Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,7 +4,7 @@ int main(){
     scanf("%d",&a);
     scanf("%d",&b);
     scanf("%d",&c);
-    float average=(a+b+c)/3.0;
+    float average=(float)(a+b+c)/3;
     printf("Average: %.2f",average);
     return 0;
